✨ 룸 리스트 새로고침 로더 연결
